Create a new module for multi_rotor to include backlash
