+++ ross/__init__.py
@@ -7,17 +7,18 @@ import ross.plotly_theme
 from .bearing_seal_element import *
 from .faults import *
 from .disk_element import *
-from .gear_element import *
 from .materials import *
 from .point_mass import *
 from .probe import *
 from .results import *
 from .rotor_assembly import *
-from .multi_rotor import *
 from .shaft_element import *
 from .coupling_element import *
 from .units import Q_
 from .utils import get_data_from_figure, visualize_matrix
+from .multi_rotor.multi_rotor import *
+from .multi_rotor.gear_element import *
+from .multi_rotor.mesh import *
 from ross.bearings.lubricants import lubricants_dict
 from ross.bearings.bearing_results import *
 from ross.bearings.plain_journal import *
